requirements.txt
Flask==3.0.3
Flask-RESTx==1.3.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytest==8.3.3
pytest-flask==1.3.0